Issue #34
Added the logic for creating an openURL with some relevant tests. This part
is essentially complete. Just need to add a few more test cases for varying
types of requests to ensure the response is that which we expect.

Currently, the object should be loaded as a mixin, rather than its stand alone
widget. It should sit inside the Resources widget, that passes it the details
such as Solr docs and the users server preference.