Issue #34
Tests added for the OpenURLGenerator within the LinkGenerator mixin.
This has placed some assumptions about the data is receives. To ensure
this is going to work, the tests for the ResourcesWidget must be
updated to ensure the correct data is passed on.

Currently, it requires two forms of data:

  - Solr 'doc' for the article
  - User preferences from the BeeHive (loggedIn and libraryLink)

Implementation of the above will require mocking from the services I
would assume - this should not be too difficult *touch wood*.