Issue #451
baseURL was not passed to the OpenURLGenerator in the LinkGeneratorMixin, meaning
all baseURLs would be of the form 'undefined'. I have added the paramter to be
passed in the LinkGeneratorMixin, and added a check in the test for the Mixin to
ensure the baseURL is set correctly this way.

ISBN was being set by a non-existent object attribute 'this.isbn' and should have
been 'this.rft_isbn'. This has been updated. The tests have been modified to check
that no 'undefined' values are found in the object, which would highlight any behaviour
like that of the ISBN.