Issue #324
This is part 2 of issue #324, which should officially close it. This addresses the behaviour
for when someone revokes their ORCID token when they have a BBBB session open at the same time.
If they were to click a widget, such as 'Claim in ORCID', it would return a 401 error, and any
widgets would remain spinning.

The 401 error is now intercepted at the discovery_mediator level, such that it checks to see
if it is an ORCID error. If it is indeed when accessing ORCID, the user is logged out of their
ORCID, and informed that they may have revoked their token - as such, they must log back in via
the ORCID panel.