Issue #570 Missing full-text links when openURL...
This is a commit for part 1) of the requirements
defined in the issue.

Modifications are such that an OpenURL is only ever
generated if the object (article, journal, whatever)
has a DOI. Otherwise, it is not made at all.

A test has been included within the link generator
mixin.

Some other tidying was done in minor places.