Add Api Query Validator
Added a new component to run some common validations on the search bar input.
Now queries that contain empty "" or "^" or ("") will be caught on the front-end,
instead of relying on the api to validate.  However, more complex queries such as
nested fields and garbled input will be passed along to the api to parse.  When
the validator catches some bad input, an error message is provided to the user.