Implement the 10k Test Suite (#214)
* Don't publish 10k test suite data yet

* Add Jython fork as a submodule

I needed to patch an issue in PyUnicode so I made a new fork; will remove if/when it gets upstreamed.

* Ignore .DS_Store files

* Add Jython submodule

* Include Jython Gradle project in Montysolr

* Add test for failing name parser test case

* Handle fully unparsable inputs in the Python code

Prior to this the code would error out because it expected at least 1 part. In cases where there are additional parenthesis around the string (it happens), and other miscellaneous inputs, there can be 0 parts instead.

* Handle unparsable author names in the Java code

This causes the unparsed author name to pass through the system. In previous versions there would be (incorrect) null characters added to the output of this pass if the author name couldn't be parsed.

* Use NFKC normalization for author names

This normalization pass helps to consolidate the Unicode code points in the string prior to other passes. Without this step some important parts of certain code points can be eliminated, resulting in mangled output.

* Add test case for parenthesized author names

* Allow `start` in position queries to be >= 0

* Add 10k test suite test case

* Un-ignore the dataset

* Add the query dataset

* Fix Jython unicode split behavior

* Downgrade Gradle to handle Jython sub-build

* Match test to new normalization code