Solr 9 Update (#221)
* Don't publish 10k test suite data yet

* Add Jython fork as a submodule

I needed to patch an issue in PyUnicode so I made a new fork; will remove if/when it gets upstreamed.

* Ignore .DS_Store files

* Add Jython submodule

* Include Jython Gradle project in Montysolr

* Add test for failing name parser test case

* Handle fully unparsable inputs in the Python code

Prior to this the code would error out because it expected at least 1 part. In cases where there are additional parenthesis around the string (it happens), and other miscellaneous inputs, there can be 0 parts instead.

* Handle unparsable author names in the Java code

This causes the unparsed author name to pass through the system. In previous versions there would be (incorrect) null characters added to the output of this pass if the author name couldn't be parsed.

* Use NFKC normalization for author names

This normalization pass helps to consolidate the Unicode code points in the string prior to other passes. Without this step some important parts of certain code points can be eliminated, resulting in mangled output.

* Add test case for parenthesized author names

* Allow `start` in position queries to be >= 0

* Add 10k test suite test case

* Un-ignore the dataset

* Add the query dataset

* Initial Solr 9 commit

Most tests are passing at this point, but I need to merge in the name normalization changes to fix issues caused by the Jython unicode split bug.

* Fix final round of Solr 9 test failures

* Upgrade to Java 17 to match Solr 9

* Match Lucene version to Solr dependency version

* Update luceneMatchVersion to fix warning

We were getting warnings about deprecated behavior at launch

* Fix locale issue in Jython with JDK 17

* Add test case to match adshelp email

* Remove lib directives that match modules

* Add test case for replication URL handling